Remove duplicate file names
